requirements.txt
streamlit==1.43.2
psycopg[binary]==3.2.6
python-dotenv==1.0.1
openai==1.66.5
pandas==2.2.3

